+++ src/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+* Implementation for mapping Duration fields. Currently handled in milliseconds
+* Implementation for mapping BigInt fields. Temporary handled as integer
+
 ## [0.6.2+1] - 2019-05-15
 
 ### Changed